+++ CHANGELOG.rst
@@ -26,6 +26,9 @@ Added
   of total function calls) are stored in the experiment.yaml file. Note that
   the logged time includes also the overhead introduced by the HDS package
   itself.
+* The following test functions are added:
+  * BreitWigner
+  * Reciprocal
 
 Changed
 -------